my answers

+++ 3/two lines.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 int main()
 {
-	float a, b, c, a2, b2, c2, x = 0, y = 0;
+	double a, b, c, a2, b2, c2, x = 0, y = 0;
 	cin >> a >> b >> c >> a2 >> b2 >> c2;
 	if (a*b2 - b*a2 == 0) {
 		cout << "No Solution";
